Ripped out Sharktopoda 1 controls
