+++ build.gradle
@@ -86,17 +86,17 @@ subprojects {
             implementation 'org.kordamp.ikonli:ikonli-swing:12.3.1'
             implementation 'org.zeromq:jeromq:0.5.4'
             implementation 'org.mbari:imgfx:0.0.15'
-            implementation 'org.mbari.vcr4j:vcr4j-core:5.2.0'
-            implementation 'org.mbari.vcr4j:vcr4j-remote:5.2.0'
-            implementation 'org.mbari.vcr4j:vcr4j-sharktopoda-client:5.2.0'
-            implementation 'org.mbari.vcr4j:vcr4j-sharktopoda:5.2.0'
+            implementation 'org.mbari.vcr4j:vcr4j-core:5.3.0'
+            implementation 'org.mbari.vcr4j:vcr4j-remote:5.3.0'
+//            implementation 'org.mbari.vcr4j:vcr4j-sharktopoda-client:5.3.0'
+//            implementation 'org.mbari.vcr4j:vcr4j-sharktopoda:5.3.0'
             implementation 'org.mbari:mbarix4j:2.0.5.jre11'
             implementation 'org.slf4j:slf4j-api:2.0.12'
             implementation 'org.slf4j:slf4j-jdk14:2.0.12'
             implementation 'org.slf4j:slf4j-jdk-platform-logging:2.0.12'
             implementation 'org.slf4j:jul-to-slf4j:2.0.12'
             implementation 'org.swinglabs.swingx:swingx-all:1.6.5-1'
-            runtimeOnly 'ch.qos.logback:logback-classic:1.5.3'
+            runtimeOnly 'ch.qos.logback:logback-classic:1.5.6'
             runtimeOnly 'org.fusesource.jansi:jansi:2.4.1'
         }
