+++ org.mbari.vars.services/src/main/java/org/mbari/vars/services/etc/gson/GsonContants.java
@@ -1,0 +1,4 @@
+package org.mbari.vars.services.etc.gson;
+
+public class GsonContants {
+}
