+++ org.mbari.vars.services/src/main/java/org/mbari/vars/services/etc/gson/TimecodeConverter.java
@@ -1,4 +1,4 @@
-package org.mbari.vars.services.gson;
+package org.mbari.vars.services.etc.gson;
 
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
