--- org.mbari.vars.services/src/main/java/org/mbari/vars/services/gson/GsonContants.java
@@ -1,4 +1,0 @@
-package org.mbari.vars.services.gson;
-
-public class GsonContants {
-}
