+++ org.mbari.vars.services/src/main/java/org/mbari/vars/services/impl/varsuserserver/v1/PrefWebServiceFactory.java
@@ -3,10 +3,10 @@ package org.mbari.vars.services.impl.varsuserserver.v1;
 //import com.fatboyindustrial.gsonjavatime.Converters;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import org.mbari.vars.services.gson.ByteArrayConverter;
-import org.mbari.vars.services.gson.DurationConverter;
-import org.mbari.vars.services.gson.InstantConverter;
-import org.mbari.vars.services.gson.TimecodeConverter;
+import org.mbari.vars.services.etc.gson.ByteArrayConverter;
+import org.mbari.vars.services.etc.gson.DurationConverter;
+import org.mbari.vars.services.etc.gson.InstantConverter;
+import org.mbari.vars.services.etc.gson.TimecodeConverter;
 import org.mbari.vars.services.RetrofitServiceFactory;
 import org.mbari.vcr4j.time.Timecode;
