+++ org.mbari.vars.ui/build.gradle
@@ -18,8 +18,8 @@ dependencies {
     implementation 'org.mbari:imgfx'
     implementation 'org.mbari.vcr4j:vcr4j-core'
     implementation 'org.mbari.vcr4j:vcr4j-remote'
-    implementation 'org.mbari.vcr4j:vcr4j-sharktopoda'
-    implementation 'org.mbari.vcr4j:vcr4j-sharktopoda-client'
+//    implementation 'org.mbari.vcr4j:vcr4j-sharktopoda'
+//    implementation 'org.mbari.vcr4j:vcr4j-sharktopoda-client'
     implementation 'org.mbari:imgfx'
     implementation 'org.mbari:mbarix4j'
     implementation 'org.slf4j:slf4j-api'
