+++ org.mbari.vars.ui/src/main/java/org/mbari/vars/ui/mediaplayers/sharktopoda/Constants.java
@@ -1,0 +1,8 @@
+package org.mbari.vars.ui.mediaplayers.sharktopoda;
+
+public class Constants {
+
+    public static final double MAX_SHUTTLE_RATE = 8.0;
+    public static final double DEFAULT_SHUTTLE_RATE = 3.0;
+    public static final Object LOCALIZATION_EVENT_SOURCE = "SharktopodaLocalizationEventSource";
+}
