+++ org.mbari.vars.ui/src/main/java/org/mbari/vars/ui/mediaplayers/sharktopoda2/IncomingController.java
@@ -7,9 +7,7 @@ import org.mbari.vars.ui.commands.CreateAnnotationAtIndexWithAssociationCmd;
 import org.mbari.vars.ui.commands.DeleteAssociationsCmd;
 import org.mbari.vars.ui.commands.UpdateAssociationCmd;
 import org.mbari.vars.ui.events.AnnotationsSelectedEvent;
-import org.mbari.vars.ui.mediaplayers.sharktopoda.localization.LocalizationController;
 import org.mbari.vcr4j.VideoIndex;
-import org.mbari.vcr4j.remote.control.RVideoIO;
 import org.mbari.vcr4j.remote.control.RemoteControl;
 import org.mbari.vcr4j.remote.control.commands.localization.*;
 import org.mbari.vcr4j.remote.player.RxControlRequestHandler;
