--- org.mbari.vars.ui/src/main/resources/fxml/LocalizationSettingsPane.fxml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.RowConstraints?>
-
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.TextField?>
-<GridPane fx:id="root" hgap="10.0" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefWidth="500.0" vgap="10.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.mbari.vars.ui.mediaplayers.sharktopoda.localization.LocalizationSettingsPaneController">
-  <columnConstraints>
-    <ColumnConstraints hgrow="NEVER" minWidth="10.0" prefWidth="100.0" />
-    <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-      <ColumnConstraints hgrow="NEVER" minWidth="10.0" prefWidth="100.0" />
-      <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-  </columnConstraints>
-  <rowConstraints>
-    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
-    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
-    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
-      <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="NEVER" />
-  </rowConstraints>
-   <children>
-      <Label text="%localization.pref.content" GridPane.columnSpan="2147483647" GridPane.vgrow="NEVER" />
-      <Label text="%localization.pref.in.port" GridPane.rowIndex="2" GridPane.vgrow="NEVER" />
-      <Label text="%localization.pref.out.port" GridPane.rowIndex="3" GridPane.vgrow="NEVER" />
-      <TextField fx:id="incomingPortTextField" GridPane.columnIndex="1" GridPane.rowIndex="2" GridPane.vgrow="NEVER" />
-      <TextField fx:id="outgoingPortTextField" GridPane.columnIndex="1" GridPane.rowIndex="3" GridPane.vgrow="NEVER" />
-      <TextField fx:id="incomingTopicTextField" minWidth="200.0" GridPane.columnIndex="3" GridPane.rowIndex="2" GridPane.vgrow="NEVER" />
-      <TextField fx:id="outgoingTopicTextField" minWidth="200.0" GridPane.columnIndex="3" GridPane.rowIndex="3" GridPane.vgrow="NEVER" />
-      <Label text="%localization.pref.in.topic" GridPane.columnIndex="2" GridPane.rowIndex="2" GridPane.vgrow="NEVER" />
-      <Label text="%localization.pref.out.topic" GridPane.columnIndex="2" GridPane.rowIndex="3" GridPane.vgrow="NEVER" />
-      <CheckBox fx:id="enabledCheckBox" text="%localization.pref.enabled" GridPane.columnSpan="2" GridPane.rowIndex="1" GridPane.vgrow="NEVER" />
-   </children>
-   <padding>
-      <Insets bottom="25.0" left="25.0" right="25.0" top="25.0" />
-   </padding>
-</GridPane>
